requirements.txt
openai
langchain
streamlit
python-dotenv
wikipedia
numexpr
